--- apps/controller/__init__.py
@@ -1,5 +1,0 @@
-# -*- codeing = utf-8 -*-
-# @Time : 2021/6/21 17:18
-# @Author : Bynn
-# @File : __init__.py.py
-# @Software: PyCharm
