+++ templates/logAnalyze.html
@@ -22,7 +22,7 @@
             </ul>
         </div>
 		<div class="center-topTitle">
-		    <h1>南网计量监控系统平台</h1>
+		    <h1>XXXX监控系统平台</h1>
 		</div>
 		<div class="right nav">
             <ul style="margin-left: 2%;">
@@ -83,8 +83,6 @@
 <script type="text/javascript">
     $('document').ready(function () {
         $("body").css('visibility', 'visible');
-        var localData = [$('#teacher').val(), $('#start').val() + '/' + $('#end').val(), $('#leader').val()]
-        localStorage.setItem("data", localData);
 
     })
 </script>
