+++ templates/suricata.html
@@ -8,7 +8,7 @@
     <script type="text/javascript" src="../static/js/rem.js"></script>
 	<script type="text/javascript" src="../static/js/jquery-3.3.1.min.js"></script>
 	<script type="text/javascript" src="../static/js/echarts.min.js"></script>
-	<script type="text/javascript" src="../static/js/china.js"></script>
+	<script type="text/javascript" src="../static/js/suricata/china.js"></script>
     <script type="text/javascript">
         var Display = {
             Data:{{ attackDisplayData|tojson }}
@@ -27,7 +27,7 @@
             </ul>
         </div>
 		<div class="center-topTitle">
-		    <h1>南网计量监控系统平台</h1>
+		    <h1>XXXX监控系统平台</h1>
 		</div>
 		<div class="right nav">
             <ul style="margin-left: 2%;">
@@ -127,9 +127,6 @@
 <script type="text/javascript">
     $('document').ready(function () {
         $("body").css('visibility', 'visible');
-        var localData = [$('#teacher').val(), $('#start').val() + '/' + $('#end').val(), $('#leader').val()]
-        localStorage.setItem("data", localData);
-
     })
 </script>
 </html>
